backend/requirements.txt
boto3==1.34.0
fastapi==0.110.0
uvicorn==0.29.0
uvicorn[standard]
pydantic==2.6.0
openai-whisper==20231117
gtts==2.5.1
pandas==2.2.0
requests==2.31.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
langchain
langchain-community
langchain-chroma
chromadb
beautifulsoup4
sentence-transformers
edge-tts
